% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ans.R
\name{ans_clear_cache}
\alias{ans_clear_cache}
\title{Clear ANS Cache}
\usage{
ans_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached ANS data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
ans_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other ans:
\code{\link[=ans_cache_status]{ans_cache_status()}},
\code{\link[=ans_data]{ans_data()}},
\code{\link[=ans_info]{ans_info()}},
\code{\link[=ans_operators]{ans_operators()}},
\code{\link[=ans_variables]{ans_variables()}},
\code{\link[=ans_years]{ans_years()}}
}
\concept{ans}
