% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anvisa.R
\name{anvisa_clear_cache}
\alias{anvisa_clear_cache}
\title{Clear ANVISA Cache}
\usage{
anvisa_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached ANVISA data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
anvisa_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other anvisa:
\code{\link[=anvisa_cache_status]{anvisa_cache_status()}},
\code{\link[=anvisa_data]{anvisa_data()}},
\code{\link[=anvisa_info]{anvisa_info()}},
\code{\link[=anvisa_types]{anvisa_types()}},
\code{\link[=anvisa_variables]{anvisa_variables()}}
}
\concept{anvisa}
