% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo.R
\name{censo_info}
\alias{censo_info}
\title{Census information}
\usage{
censo_info(year = NULL)
}
\arguments{
\item{year}{Numeric. Year to get specific information about.
NULL shows general info.}
}
\value{
Invisibly returns a list with Census metadata.
}
\description{
Displays information about the Brazilian Census and returns metadata.
}
\examples{
censo_info()
censo_info(2022)
}
