% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo.R
\name{censo_years}
\alias{censo_years}
\title{List available Census years}
\usage{
censo_years()
}
\value{
A character vector of available years.
}
\description{
Returns a character vector with available Census years.
}
\examples{
censo_years()
}
