% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{create_partitioned_cache}
\alias{create_partitioned_cache}
\title{Create partitioned parquet cache}
\usage{
create_partitioned_cache(df, cache_dir)
}
\arguments{
\item{df}{A data frame with VIGITEL data}

\item{cache_dir}{Character. Cache directory path.}
}
\value{
Invisible path to the parquet directory, or NULL if arrow not available.
}
\description{
Create partitioned parquet cache
}
\keyword{internal}
