% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{get_arrow_column_names}
\alias{get_arrow_column_names}
\title{Get all available columns from Arrow dataset}
\usage{
get_arrow_column_names(dataset)
}
\arguments{
\item{dataset}{An Arrow Dataset}
}
\value{
Character vector of column names.
}
\description{
Get all available columns from Arrow dataset
}
\keyword{internal}
