% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{has_arrow}
\alias{has_arrow}
\title{Check if arrow package is available}
\usage{
has_arrow()
}
\value{
TRUE if arrow is available, FALSE otherwise
}
\description{
Check if arrow package is available
}
\keyword{internal}
