% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{has_partitioned_cache}
\alias{has_partitioned_cache}
\title{Check if partitioned cache exists}
\usage{
has_partitioned_cache(cache_dir)
}
\arguments{
\item{cache_dir}{Character. Cache directory path.}
}
\value{
Logical. TRUE if partitioned cache exists.
}
\description{
Check if partitioned cache exists
}
\keyword{internal}
