% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc.R
\name{pnadc_info}
\alias{pnadc_info}
\title{PNADC survey information}
\usage{
pnadc_info()
}
\value{
Invisibly returns a list with survey metadata.
}
\description{
Displays information about PNAD Continua and returns metadata.
}
\examples{
pnadc_info()
}
