% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc.R
\name{pnadc_modules}
\alias{pnadc_modules}
\title{List available PNADC modules}
\usage{
pnadc_modules()
}
\value{
A tibble with module information including name, available years,
and descriptions.
}
\description{
Returns information about the available supplementary modules in PNAD Continua
that are supported by this package.
}
\examples{
pnadc_modules()
}
