% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc.R
\name{pnadc_years}
\alias{pnadc_years}
\title{List available years for a PNADC module}
\usage{
pnadc_years(module)
}
\arguments{
\item{module}{Character. The module identifier. Use \code{\link{pnadc_modules}}
to see available modules.}
}
\value{
An integer vector of available years.
}
\description{
Returns a vector of years for which data is available for the specified module.
}
\examples{
pnadc_years("deficiencia")
pnadc_years("habitacao")
}
