% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pns.R
\name{pns_info}
\alias{pns_info}
\title{PNS survey information}
\usage{
pns_info(year = NULL)
}
\arguments{
\item{year}{Numeric. Year to get specific information about.
NULL shows general info.}
}
\value{
Invisibly returns a list with survey metadata.
}
\description{
Displays information about the PNS survey and returns metadata.
}
\examples{
pns_info()
pns_info(2019)
}
