% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pns.R
\name{pns_modules}
\alias{pns_modules}
\title{List PNS survey modules}
\usage{
pns_modules(year = NULL)
}
\arguments{
\item{year}{Numeric. Year to get modules for (2013 or 2019).
NULL returns modules for all years. Default is NULL.}
}
\value{
A tibble with module codes, names, and descriptions.
}
\description{
Returns information about the questionnaire modules available in the PNS.
}
\examples{
pns_modules()
pns_modules(year = 2019)
}
