% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pns.R
\name{pns_years}
\alias{pns_years}
\title{List available PNS survey years}
\usage{
pns_years()
}
\value{
A character vector of available years.
}
\description{
Returns a character vector with available PNS survey years.
}
\examples{
pns_years()
}
