% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sia.R
\name{sia_clear_cache}
\alias{sia_clear_cache}
\title{Clear SIA Cache}
\usage{
sia_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached SIA data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
sia_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other sia:
\code{\link[=sia_cache_status]{sia_cache_status()}},
\code{\link[=sia_data]{sia_data()}},
\code{\link[=sia_dictionary]{sia_dictionary()}},
\code{\link[=sia_info]{sia_info()}},
\code{\link[=sia_variables]{sia_variables()}},
\code{\link[=sia_years]{sia_years()}}
}
\concept{sia}
