% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_clear_cache}
\alias{sim_clear_cache}
\title{Clear SIM Cache}
\usage{
sim_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached SIM data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
sim_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other sim:
\code{\link[=sim_cache_status]{sim_cache_status()}},
\code{\link[=sim_data]{sim_data()}},
\code{\link[=sim_dictionary]{sim_dictionary()}},
\code{\link[=sim_info]{sim_info()}},
\code{\link[=sim_variables]{sim_variables()}},
\code{\link[=sim_years]{sim_years()}}
}
\concept{sim}
