% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinan.R
\name{sinan_cache_status}
\alias{sinan_cache_status}
\title{Show SINAN Cache Status}
\usage{
sinan_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached SINAN data files.
}
\examples{
sinan_cache_status()
}
\seealso{
Other sinan:
\code{\link[=sinan_clear_cache]{sinan_clear_cache()}},
\code{\link[=sinan_data]{sinan_data()}},
\code{\link[=sinan_dictionary]{sinan_dictionary()}},
\code{\link[=sinan_diseases]{sinan_diseases()}},
\code{\link[=sinan_info]{sinan_info()}},
\code{\link[=sinan_variables]{sinan_variables()}},
\code{\link[=sinan_years]{sinan_years()}}
}
\concept{sinan}
