% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_convert_to_parquet}
\alias{vigitel_convert_to_parquet}
\title{Convert Excel file to Parquet format}
\usage{
vigitel_convert_to_parquet(year, force = FALSE)
}
\arguments{
\item{year}{Integer year}

\item{force}{Logical. If TRUE, reconvert even if parquet exists.}
}
\value{
Path to parquet file (invisibly)
}
\description{
Convert Excel file to Parquet format
}
\keyword{internal}
