% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_download_data}
\alias{vigitel_download_data}
\title{Download VIGITEL data file}
\usage{
vigitel_download_data(format, destfile)
}
\arguments{
\item{format}{Character. "dta" or "csv".}

\item{destfile}{Character. Destination path for the ZIP file.}
}
\value{
Invisible NULL. Called for side effect (file download).
}
\description{
Download VIGITEL data file
}
\keyword{internal}
