% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_excel_path}
\alias{vigitel_excel_path}
\title{Get path to Excel file for a specific year}
\usage{
vigitel_excel_path(year, cache_dir = NULL)
}
\arguments{
\item{year}{Integer year}

\item{cache_dir}{Optional custom cache directory}
}
\value{
Path to excel file
}
\description{
Get path to Excel file for a specific year
}
\keyword{internal}
