% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_extract_zip}
\alias{vigitel_extract_zip}
\title{Extract VIGITEL ZIP file}
\usage{
vigitel_extract_zip(zip_path, exdir)
}
\arguments{
\item{zip_path}{Character. Path to the ZIP file.}

\item{exdir}{Character. Directory to extract to.}
}
\value{
Character. Path to the extracted file.
}
\description{
Extract VIGITEL ZIP file
}
\keyword{internal}
