% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_file_url}
\alias{vigitel_file_url}
\title{Build VIGITEL file URL for a specific year}
\usage{
vigitel_file_url(year)
}
\arguments{
\item{year}{Integer year}
}
\value{
Character string with file URL
}
\description{
Build VIGITEL file URL for a specific year
}
\keyword{internal}
