% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_identify_year_column}
\alias{vigitel_identify_year_column}
\title{Identify year column in VIGITEL data}
\usage{
vigitel_identify_year_column(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Character. Name of the year column.
}
\description{
Identify year column in VIGITEL data
}
\keyword{internal}
