% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_identify_year_column_from_schema}
\alias{vigitel_identify_year_column_from_schema}
\title{Identify year column from Arrow schema}
\usage{
vigitel_identify_year_column_from_schema(dataset)
}
\arguments{
\item{dataset}{An Arrow Dataset}
}
\value{
Character. Name of the year column.
}
\description{
Identify year column from Arrow schema
}
\keyword{internal}
