% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_info}
\alias{vigitel_info}
\title{Get VIGITEL survey information}
\usage{
vigitel_info()
}
\value{
A list with survey information
}
\description{
Returns metadata about the VIGITEL survey.
}
\examples{
vigitel_info()
}
