% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_parquet_path}
\alias{vigitel_parquet_path}
\title{Get path to Parquet file for a specific year}
\usage{
vigitel_parquet_path(year)
}
\arguments{
\item{year}{Integer year}
}
\value{
Path to parquet file
}
\description{
Get path to Parquet file for a specific year
}
\keyword{internal}
