% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_parse_years}
\alias{vigitel_parse_years}
\title{Parse year argument}
\usage{
vigitel_parse_years(year)
}
\arguments{
\item{year}{Year specification (integer, character, vector, or "all")}
}
\value{
Integer vector of years
}
\description{
Converts various year input formats to integer vector.
}
\keyword{internal}
