% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_read_data}
\alias{vigitel_read_data}
\title{Read VIGITEL data file}
\usage{
vigitel_read_data(path, format)
}
\arguments{
\item{path}{Character. Path to the data file (.dta or .csv).}

\item{format}{Character. "dta" or "csv".}
}
\value{
A tibble.
}
\description{
Read VIGITEL data file
}
\keyword{internal}
