% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_years}
\alias{vigitel_years}
\title{List available VIGITEL survey years}
\usage{
vigitel_years()
}
\value{
An integer vector of available years
}
\description{
Returns a vector of years for which VIGITEL microdata is available
for download from the Ministry of Health website.
}
\examples{
vigitel_years()
}
