% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ans.R
\name{ans_cache_status}
\alias{ans_cache_status}
\title{Show ANS Cache Status}
\usage{
ans_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached ANS data files.
}
\examples{
ans_cache_status()
}
\seealso{
Other ans: 
\code{\link{ans_clear_cache}()},
\code{\link{ans_data}()},
\code{\link{ans_info}()},
\code{\link{ans_operators}()},
\code{\link{ans_variables}()},
\code{\link{ans_years}()}
}
\concept{ans}
