% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ans.R
\name{ans_info}
\alias{ans_info}
\title{ANS Module Information}
\usage{
ans_info()
}
\value{
A list with module information (invisibly).
}
\description{
Displays information about the ANS (Agencia Nacional de Saude Suplementar)
module, including data sources, available years, and usage guidance.
}
\examples{
ans_info()
}
\seealso{
Other ans: 
\code{\link{ans_cache_status}()},
\code{\link{ans_clear_cache}()},
\code{\link{ans_data}()},
\code{\link{ans_operators}()},
\code{\link{ans_variables}()},
\code{\link{ans_years}()}
}
\concept{ans}
