% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnes.R
\name{cnes_cache_status}
\alias{cnes_cache_status}
\title{Show CNES Cache Status}
\usage{
cnes_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached CNES data files.
}
\examples{
cnes_cache_status()
}
\seealso{
Other cnes: 
\code{\link{cnes_clear_cache}()},
\code{\link{cnes_data}()},
\code{\link{cnes_dictionary}()},
\code{\link{cnes_info}()},
\code{\link{cnes_variables}()},
\code{\link{cnes_years}()}
}
\concept{cnes}
