% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnes.R
\name{cnes_clear_cache}
\alias{cnes_clear_cache}
\title{Clear CNES Cache}
\usage{
cnes_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached CNES data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
cnes_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other cnes: 
\code{\link{cnes_cache_status}()},
\code{\link{cnes_data}()},
\code{\link{cnes_dictionary}()},
\code{\link{cnes_info}()},
\code{\link{cnes_variables}()},
\code{\link{cnes_years}()}
}
\concept{cnes}
