% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthbR-package.R
\docType{package}
\name{healthbR-package}
\alias{healthbR}
\alias{healthbR-package}
\title{healthbR: Access Brazilian Public Health Data}
\description{
Provides easy access to Brazilian public health data from multiple sources including VIGITEL (Surveillance of Risk Factors for Chronic Diseases by Telephone Survey), PNS (National Health Survey), SIM (Mortality Information System), SINASC (Live Birth Information System), and other health information systems. Data is downloaded from the Brazilian Ministry of Health VIGITEL repository \url{https://svs.aids.gov.br/download/Vigitel/}. Data is returned in tidy format following tidyverse conventions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SidneyBissoli/healthbR}
  \item Report bugs at \url{https://github.com/SidneyBissoli/healthbR/issues}
}

}
\author{
\strong{Maintainer}: Sidney Bissoli \email{sbissoli76@gmail.com} (\href{https://orcid.org/0009-0001-0442-3700}{ORCID})

}
\keyword{internal}
