% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthbR-package.R
\docType{package}
\name{healthbR-package}
\alias{healthbR}
\alias{healthbR-package}
\title{healthbR: Access Brazilian Public Health Data}
\description{
Provides easy access to Brazilian public health data from multiple sources including VIGITEL (Surveillance of Risk Factors for Chronic Diseases by Telephone Survey), PNS (National Health Survey), 'PNAD' Continua (Continuous National Household Sample Survey), 'POF' (Household Budget Survey with food security and consumption data), 'Censo Demografico' (population denominators via 'SIDRA' API), SIM (Mortality Information System), SINASC (Live Birth Information System), 'SIH' (Hospital Information System), 'SIA' (Outpatient Information System), 'SINAN' (Notifiable Diseases Surveillance), 'CNES' (National Health Facility Registry), 'SI-PNI' (National Immunization Program - aggregated 1994-2019 via FTP, individual-level 'microdata' 2020+ via 'OpenDataSUS' API), 'SISAB' (Primary Care Health Information System - coverage indicators via REST API), ANS ('Agencia Nacional de Saude Suplementar' - supplementary health beneficiaries, consumer complaints, and financial statements), 'ANVISA' ('Agencia Nacional de Vigilancia Sanitaria' - product registrations, 'pharmacovigilance', 'hemovigilance', 'technovigilance', and controlled substance sales via 'SNGPC'), and other health information systems. Data is downloaded from the Brazilian Ministry of Health and 'IBGE' repositories. Data is returned in tidy format following tidyverse conventions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SidneyBissoli/healthbR}
  \item \url{https://sidneybissoli.github.io/healthbR/}
  \item Report bugs at \url{https://github.com/SidneyBissoli/healthbR/issues}
}

}
\author{
\strong{Maintainer}: Sidney Bissoli \email{sbissoli76@gmail.com} (\href{https://orcid.org/0009-0001-0442-3700}{ORCID})

}
\keyword{internal}
