% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof.R
\name{pof_clear_cache}
\alias{pof_clear_cache}
\title{Clear POF cache}
\usage{
pof_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Optional custom cache directory. If NULL (default),
uses the standard user cache directory.}
}
\value{
NULL (invisibly)
}
\description{
Removes all cached POF data files.
}
\examples{
pof_clear_cache()
}
\seealso{
Other pof: 
\code{\link{pof_cache_status}()},
\code{\link{pof_data}()},
\code{\link{pof_dictionary}()},
\code{\link{pof_info}()},
\code{\link{pof_registers}()},
\code{\link{pof_variables}()},
\code{\link{pof_years}()}
}
\concept{pof}
