% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof.R
\name{pof_years}
\alias{pof_years}
\title{List available POF survey years}
\usage{
pof_years()
}
\value{
A character vector of available years in "YYYY-YYYY" format.
}
\description{
Returns a character vector with available POF survey years.
}
\examples{
pof_years()
}
\seealso{
Other pof: 
\code{\link{pof_cache_status}()},
\code{\link{pof_clear_cache}()},
\code{\link{pof_data}()},
\code{\link{pof_dictionary}()},
\code{\link{pof_info}()},
\code{\link{pof_registers}()},
\code{\link{pof_variables}()}
}
\concept{pof}
