% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sia.R
\name{sia_cache_status}
\alias{sia_cache_status}
\title{Show SIA Cache Status}
\usage{
sia_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached SIA data files.
}
\examples{
sia_cache_status()
}
\seealso{
Other sia: 
\code{\link{sia_clear_cache}()},
\code{\link{sia_data}()},
\code{\link{sia_dictionary}()},
\code{\link{sia_info}()},
\code{\link{sia_variables}()},
\code{\link{sia_years}()}
}
\concept{sia}
