% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sih.R
\name{sih_cache_status}
\alias{sih_cache_status}
\title{Show SIH Cache Status}
\usage{
sih_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached SIH data files.
}
\examples{
sih_cache_status()
}
\seealso{
Other sih: 
\code{\link{sih_clear_cache}()},
\code{\link{sih_data}()},
\code{\link{sih_dictionary}()},
\code{\link{sih_info}()},
\code{\link{sih_variables}()},
\code{\link{sih_years}()}
}
\concept{sih}
