% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sih.R
\name{sih_clear_cache}
\alias{sih_clear_cache}
\title{Clear SIH Cache}
\usage{
sih_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached SIH data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
sih_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other sih: 
\code{\link{sih_cache_status}()},
\code{\link{sih_data}()},
\code{\link{sih_dictionary}()},
\code{\link{sih_info}()},
\code{\link{sih_variables}()},
\code{\link{sih_years}()}
}
\concept{sih}
