% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinan.R
\name{sinan_clear_cache}
\alias{sinan_clear_cache}
\title{Clear SINAN Cache}
\usage{
sinan_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached SINAN data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
sinan_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other sinan: 
\code{\link{sinan_cache_status}()},
\code{\link{sinan_data}()},
\code{\link{sinan_dictionary}()},
\code{\link{sinan_diseases}()},
\code{\link{sinan_info}()},
\code{\link{sinan_variables}()},
\code{\link{sinan_years}()}
}
\concept{sinan}
