% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinan.R
\name{sinan_info}
\alias{sinan_info}
\title{SINAN Module Information}
\usage{
sinan_info()
}
\value{
A list with module information (invisibly).
}
\description{
Displays information about the Notifiable Diseases Information System (SINAN),
including data sources, available years, diseases, and usage guidance.
}
\examples{
sinan_info()
}
\seealso{
Other sinan: 
\code{\link{sinan_cache_status}()},
\code{\link{sinan_clear_cache}()},
\code{\link{sinan_data}()},
\code{\link{sinan_dictionary}()},
\code{\link{sinan_diseases}()},
\code{\link{sinan_variables}()},
\code{\link{sinan_years}()}
}
\concept{sinan}
