% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinasc.R
\name{sinasc_cache_status}
\alias{sinasc_cache_status}
\title{Show SINASC Cache Status}
\usage{
sinasc_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached SINASC data files.
}
\examples{
sinasc_cache_status()
}
\seealso{
Other sinasc: 
\code{\link{sinasc_clear_cache}()},
\code{\link{sinasc_data}()},
\code{\link{sinasc_dictionary}()},
\code{\link{sinasc_info}()},
\code{\link{sinasc_variables}()},
\code{\link{sinasc_years}()}
}
\concept{sinasc}
