% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinasc.R
\name{sinasc_clear_cache}
\alias{sinasc_clear_cache}
\title{Clear SINASC Cache}
\usage{
sinasc_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached SINASC data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
sinasc_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other sinasc: 
\code{\link{sinasc_cache_status}()},
\code{\link{sinasc_data}()},
\code{\link{sinasc_dictionary}()},
\code{\link{sinasc_info}()},
\code{\link{sinasc_variables}()},
\code{\link{sinasc_years}()}
}
\concept{sinasc}
