% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipni.R
\name{sipni_cache_status}
\alias{sipni_cache_status}
\title{Show SI-PNI Cache Status}
\usage{
sipni_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached SI-PNI data files.
}
\examples{
sipni_cache_status()
}
\seealso{
Other sipni: 
\code{\link{sipni_clear_cache}()},
\code{\link{sipni_data}()},
\code{\link{sipni_dictionary}()},
\code{\link{sipni_info}()},
\code{\link{sipni_variables}()},
\code{\link{sipni_years}()}
}
\concept{sipni}
