% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipni.R
\name{sipni_clear_cache}
\alias{sipni_clear_cache}
\title{Clear SI-PNI Cache}
\usage{
sipni_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached SI-PNI data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
sipni_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other sipni: 
\code{\link{sipni_cache_status}()},
\code{\link{sipni_data}()},
\code{\link{sipni_dictionary}()},
\code{\link{sipni_info}()},
\code{\link{sipni_variables}()},
\code{\link{sipni_years}()}
}
\concept{sipni}
