% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipni.R
\name{sipni_info}
\alias{sipni_info}
\title{SI-PNI Module Information}
\usage{
sipni_info()
}
\value{
A list with module information (invisibly).
}
\description{
Displays information about the National Immunization Program Information
System (SI-PNI), including data sources, available years, file types,
and usage guidance.
}
\examples{
sipni_info()
}
\seealso{
Other sipni: 
\code{\link{sipni_cache_status}()},
\code{\link{sipni_clear_cache}()},
\code{\link{sipni_data}()},
\code{\link{sipni_dictionary}()},
\code{\link{sipni_variables}()},
\code{\link{sipni_years}()}
}
\concept{sipni}
