% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipni.R
\name{sipni_years}
\alias{sipni_years}
\title{List Available SI-PNI Years}
\usage{
sipni_years()
}
\value{
An integer vector of available years (1994--2019).
}
\description{
Returns an integer vector with years for which vaccination data are
available from DATASUS FTP.
}
\details{
SI-PNI data on the DATASUS FTP is available from 1994 to 2019.
All data is definitive (no preliminary/final distinction).
Post-2019 data requires the SI-PNI web API (not yet supported).
}
\examples{
sipni_years()
}
\seealso{
Other sipni: 
\code{\link{sipni_cache_status}()},
\code{\link{sipni_clear_cache}()},
\code{\link{sipni_data}()},
\code{\link{sipni_dictionary}()},
\code{\link{sipni_info}()},
\code{\link{sipni_variables}()}
}
\concept{sipni}
