% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisab.R
\name{sisab_cache_status}
\alias{sisab_cache_status}
\title{Show SISAB Cache Status}
\usage{
sisab_cache_status(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
A tibble with cache file information (invisibly).
}
\description{
Shows information about cached SISAB data files.
}
\examples{
sisab_cache_status()
}
\seealso{
Other sisab: 
\code{\link{sisab_clear_cache}()},
\code{\link{sisab_data}()},
\code{\link{sisab_info}()},
\code{\link{sisab_variables}()},
\code{\link{sisab_years}()}
}
\concept{sisab}
