% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisab.R
\name{sisab_clear_cache}
\alias{sisab_clear_cache}
\title{Clear SISAB Cache}
\usage{
sisab_clear_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Cache directory path. Default:
\code{tools::R_user_dir("healthbR", "cache")}.}
}
\value{
Invisible NULL.
}
\description{
Deletes cached SISAB data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
sisab_clear_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Other sisab:
\code{\link[=sisab_cache_status]{sisab_cache_status()}},
\code{\link[=sisab_data]{sisab_data()}},
\code{\link[=sisab_info]{sisab_info()}},
\code{\link[=sisab_variables]{sisab_variables()}},
\code{\link[=sisab_years]{sisab_years()}}
}
\concept{sisab}
