% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_cache_dir}
\alias{vigitel_cache_dir}
\title{Get VIGITEL cache directory}
\usage{
vigitel_cache_dir()
}
\value{
Path to cache directory
}
\description{
Get VIGITEL cache directory
}
\keyword{internal}
