% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_cache_status}
\alias{vigitel_cache_status}
\title{Get VIGITEL cache status}
\usage{
vigitel_cache_status()
}
\value{
A tibble with cache information
}
\description{
Shows which years are cached and file sizes.
}
\examples{
\donttest{
vigitel_cache_status()
}
}
