% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigitel.R
\name{vigitel_variables}
\alias{vigitel_variables}
\title{List VIGITEL variables}
\usage{
vigitel_variables(year)
}
\arguments{
\item{year}{Integer. Year of the survey.}
}
\value{
A character vector of variable names
}
\description{
Returns a character vector of variable names available in a VIGITEL
survey year.
}
\examples{
\donttest{
# list variables for 2023
vigitel_variables(2023)
}
}
